pyproject.toml
[tool.poetry]
name = "commercial-network"
version = "0.1.0"
description = ""
authors = ["Alexander Perov <perov97@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
flake8 = "^7.0.0"
coverage = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
